requirements.txt
Django==1.11.10
django-rest-framework==0.1.0
djangorestframework==3.7.7
drf-dynamic-fields==0.2.0
pytz==2017.3
